Implement complete event sourcing booking system with consistent i64 IDs
Event Sourcing Architecture:
- Add Event and BookingCreatedEvent types with tagged union serialization
- Implement EventProcessor for transactional event storage and projection updates
- Create projections module for handling booking creation events
- Add AppState for shared database and event processor access

Database Schema Updates:
- Convert all IDs to BIGINT/i64 for consistency (hotels, bookings, events)
- Add booking_id_seq PostgreSQL sequence for atomic ID generation
- Update bookings table to use explicit BIGINT primary key instead of SERIAL
- Modify events table to use BIGSERIAL for auto-incrementing event IDs

API Implementation:
- Add CreateBookingRequest model for partial booking data input
- Implement POST /hotels/:id/bookings endpoint with full event sourcing flow
- Remove unused Room model and get_rooms handler (not implemented)
- Update all handlers to use i64 for hotel and booking IDs

Key Features:
- Atomic transactions ensuring event storage and projection consistency
- PostgreSQL sequence-based ID generation preventing race conditions
- Type-safe event serialization with serde tagged unions
- Comprehensive error handling for database operations and event processing
- CORS support and proper HTTP status codes

Dependencies:
- Switch sqlx from runtime-tokio-rustls to runtime-tokio with tls-native-tls

+++ backend/migrations/001_create_events_table.sql
@@ -2,9 +2,9 @@
 -- Based on: https://softwaremill.com/implementing-event-sourcing-using-a-relational-database/
 
 CREATE TABLE events (
-    id        SERIAL PRIMARY KEY,
+    id        BIGSERIAL PRIMARY KEY,
     stream_id BIGINT NOT NULL,
-    version   BIGINT NOT NULL,
+    version   INTEGER NOT NULL,
     data      JSONB  NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
